requirements.txt

langchain 
langchain_groq 
langchain_community
langgraph 
langsmith

langchain-openai

streamlit 

python-dotenv
